RDMA/rtrs: Add enable_folio to control folio in rtrs
Folio is a new technology in memory management. It brings a lot of
ideas in mmu. Currently folio is implemented in rtrs. In this
implementation, folio can work as DMA to interact with RDMA HW device.
In the past, folio is mostly implemented in file system. Now in RDMA,
folio can still work well.And after long time performance tests, folio
can make the whole rtrs work well. That is, with folio, the performance
remains unchanged.
Because folio is still in development, by default, folio is disabled in
rtrs server.

Signed-off-by: Zhu Yanjun <yanjun.zhu@linux.dev>